Added full_text field to data and finalised library scripts to do it; there may be the need to have more than one field, each with its relevance indicator. Fixed some bugs. Identified a PHP bug in Service.php: aggregation pipeline is not accepted although correct.

+++ Library/OntologyWrapper/Node.php
@@ -1053,10 +1053,11 @@ class Node extends MetadataObject
 			//
 			$this->offsetSet(
 				kTAG_NID,
-				static::ResolveCollection(
-					static::ResolveDatabase( $this->mDictionary, TRUE ) )
-						->getSequenceNumber(
-							static::kSEQ_NAME ) );
+				(int)
+					static::ResolveCollection(
+						static::ResolveDatabase( $this->mDictionary, TRUE ) )
+							->getSequenceNumber(
+								static::kSEQ_NAME ) );
 		
 		} // Not committed.
